+++ include/GafferScene/Filter.h
@@ -74,7 +74,7 @@ class GAFFERSCENE_API Filter : public Gaffer::ComputeNode
 		/// > where `input` is a child of a ScenePlug that will later be provided to `computeMatch()`.
 		void affects( const Gaffer::Plug *input, AffectedPlugsContainer &outputs ) const override;
 
-		/// \deprecated Use FilterPlug::SceneScope instead.
+		/// \deprecated Use FilterPlug::SceneScope or FilterPlug::match instead.
 		static void setInputScene( Gaffer::Context *context, const ScenePlug *scenePlug );
 		/// \deprecated
 		static const ScenePlug *getInputScene( const Gaffer::Context *context );