+++ Changes.md
@@ -1,6 +1,11 @@
 0.58.6.x (relative to 0.58.6.0)
 ========
 
+Fixes
+-----
+
+- Viewer : Fixed bug that caused the Inspector to edit the wrong node when SetFilters were in use.
+
 API
 ---